git-svn-id: svn://svn.code.sf.net/p/daetools/code@81 e944343b-51ee-4bc0-bd0d-d7880edd40ba

+++ trunk/LA_SuperLU/superlu_la_solver.h
@@ -145,7 +145,7 @@ public:
     int*				m_perm_r;
 #endif
 	
-	uint64_t m_solve, m_factorize;
+	double m_solve, m_factorize;
 };
 
 }
